Switched out all cuda device stuff for mps and/or cpu device, super slow though and not sure if its working yet

+++ fid.py
@@ -57,15 +57,15 @@ def calculate_inception_stats(
 
     # Accumulate statistics.
     dist.print0(f'Calculating statistics for {len(dataset_obj)} images...')
-    mu = torch.zeros([feature_dim], dtype=torch.float64, device=device)
-    sigma = torch.zeros([feature_dim, feature_dim], dtype=torch.float64, device=device)
+    mu = torch.zeros([feature_dim], dtype=torch.float32, device=device)
+    sigma = torch.zeros([feature_dim, feature_dim], dtype=torch.float32, device=device)
     for images, _labels in tqdm.tqdm(data_loader, unit='batch', disable=(dist.get_rank() != 0)):
         torch.distributed.barrier()
         if images.shape[0] == 0:
             continue
         if images.shape[1] == 1:
             images = images.repeat([1, 3, 1, 1])
-        features = detector_net(images.to(device), **detector_kwargs).to(torch.float64)
+        features = detector_net(images.to(device), **detector_kwargs).to(torch.float32)
         mu += features.sum(0)
         sigma += features.T @ features
